libcamera: Use DEP5 to specify license of files that don't support SPDX
Some files can't use SPDX as they don't support comments. This is the
case of the JSON files used by the Raspberry Pi IPA and camera tuning
tools, as well as the PGM reference picture in the camera tuning tool.

The REUSE specification allows handling this case in two different ways:

- Adding a ${file}.license file containing the license specification
  alongside each file.

- Adding a .reuse/dep5 file, in Debian machine-readable copyright format
  (as specified in
  https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/).

The first option isn't practical, use the latter.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>