qcam: dng_writer: Record creation time in the EXIF directory
If the EXIF directory is empty due to no metadata being available tools
such as tiffinfo complains that the directory is malformed.

    TIFFFetchDirectory: Sanity check on directory count failed, this is probably not a valid IFD offset.
    TIFFReadCustomDirectory: Failed to read custom directory at offset 0.

Always record the creation time in the EXIF directory instead of adding
complexity to skip creating the EXIF directory if there is no metadata
to record. This ensures there are at least some entries in the EXIF
directory and that makes tiffinfo happy.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>